Modified SQL Import file for edcon setup...

+++ dotfiles/magento/custom_import.sql
@@ -1,6 +1,6 @@
-CREATE DATABASE IF NOT EXISTS {{db_name}} ;
-DROP DATABASE IF EXISTS {{db_name}} ;
-CREATE DATABASE {{db_name}} ;
+CREATE DATABASE IF NOT EXISTS #db_name# ;
+DROP DATABASE IF EXISTS #db_name# ;
+CREATE DATABASE #db_name# ;
 \! echo "Created DB Successfully!";
 
 \! echo "Applying fixes for trigger related issues...";
@@ -8,40 +8,40 @@ CREATE USER IF NOT EXISTS 'root'@'%';
 GRANT ALL ON *.* TO 'root'@'%';
 
 
-USE {{db_name}} ;
+USE #db_name# ;
 
 \! echo "Some modifications to the SQL statement..."
 \! echo "Changing the DEFINER to the specified mysql user..."
-\! sed -i.bak -e 's/\(DEFINER=\`\)\([a-zA-Z0-9]*\)\(\`@\`\)/\1{{mysql_username}}\3/g' {{backup_sql_path}};
+\! sed -i.bak -e 's/\(DEFINER=\`\)\([a-zA-Z0-9]*\)\(\`@\`\)/\1#mysql_username#\3/g' #backup_sql_path#;
 
 \! echo "Importing DB from backup...!";
-source {{backup_sql_path}};
+source #backup_sql_path#;
 \! echo "Imported Successfully!";
 
-\! echo "Cleaning up {{db_name}} database!";
-DELETE FROM `{{db_name}}`.adminnotification_inbox;
+\! echo "Cleaning up #db_name# database!";
+DELETE FROM `#db_name#`.adminnotification_inbox;
 
-\! echo "Creating admin user for {{db_name}}...";
-UPDATE `{{db_name}}`.core_cache_option SET value = 1;
-UPDATE `{{db_name}}`.core_config_data SET value = '{{full_url}}' WHERE path = 'web/unsecure/base_url';
-UPDATE `{{db_name}}`.core_config_data SET value = '{{full_url}}' WHERE path = 'web/secure/base_url';
-UPDATE `{{db_name}}`.core_config_data SET value = '{{cookie_url}}' WHERE path = 'web/cookie/cookie_domain';
-UPDATE `{{db_name}}`.core_config_data SET value = '{{google_tag_id}}' WHERE path = 'googletagmanager/general_settings/tag_id';
-UPDATE `{{db_name}}`.core_config_data SET VALUE = 1 WHERE path = "payment/banktransfer/active";
-DELETE FROM `{{db_name}}`.`admin_role` WHERE user_id IN (SELECT user_id FROM `{{db_name}}`.`admin_user` WHERE username  = "admin");
-DELETE FROM `{{db_name}}`.`admin_user` WHERE username = "admin";
+\! echo "Creating admin user for #db_name#...";
+UPDATE `#db_name#`.core_cache_option SET value = 1;
+UPDATE `#db_name#`.core_config_data SET value = '#full_url#' WHERE path = 'web/unsecure/base_url';
+UPDATE `#db_name#`.core_config_data SET value = '#full_url#' WHERE path = 'web/secure/base_url';
+UPDATE `#db_name#`.core_config_data SET value = '#cookie_url#' WHERE path = 'web/cookie/cookie_domain';
+UPDATE `#db_name#`.core_config_data SET value = '#google_tag_id#' WHERE path = 'googletagmanager/general_settings/tag_id';
+UPDATE `#db_name#`.core_config_data SET VALUE = 1 WHERE path = "payment/banktransfer/active";
+DELETE FROM `#db_name#`.`admin_role` WHERE user_id IN (SELECT user_id FROM `#db_name#`.`admin_user` WHERE username  = "admin");
+DELETE FROM `#db_name#`.`admin_user` WHERE username = "admin";
 
-LOCK TABLES `{{db_name}}`.`admin_role` WRITE , `{{db_name}}`.`admin_user` WRITE;
+LOCK TABLES `#db_name#`.`admin_role` WRITE , `#db_name#`.`admin_user` WRITE;
 
 SET @SALT = "rp";
 SET @PASS = CONCAT(MD5(CONCAT( @SALT , "admin123") ), CONCAT(":", @SALT ));
 SELECT @EXTRA := MAX(extra) FROM admin_user WHERE extra IS NOT NULL;
 
-INSERT INTO `{{db_name}}`.`admin_user` (firstname,lastname,email,username,password,created,lognum,reload_acl_flag,is_active,extra,rp_token_created_at)
-VALUES ('{{first_name}}','{{last_name}}','{{email}}','admin',@PASS,NOW(),0,0,1,@EXTRA,NOW());
+INSERT INTO `#db_name#`.`admin_user` (firstname,lastname,email,username,password,created,lognum,reload_acl_flag,is_active,extra,rp_token_created_at)
+VALUES ('#first_name#','#last_name#','#email#','admin',@PASS,NOW(),0,0,1,@EXTRA,NOW());
 
-INSERT INTO `{{db_name}}`.`admin_role` (parent_id,tree_level,sort_order,role_type,user_id,role_name)
-VALUES (1,2,0,'U',(SELECT user_id FROM admin_user WHERE username = 'admin'),'{{first_name}}');
+INSERT INTO `#db_name#`.`admin_role` (parent_id,tree_level,sort_order,role_type,user_id,role_name)
+VALUES (1,2,0,'U',(SELECT user_id FROM admin_user WHERE username = 'admin'),'#first_name#');
 
 UNLOCK TABLES;
 
@@ -207,8 +207,8 @@ SET FOREIGN_KEY_CHECKS=1;
 
 \! echo "Clean Up completed Successfully!";
 
-\! echo "Backing up the {{db_name}} database...";
+\! echo "Backing up the #db_name# database...";
 \! echo "You will need to enter in your system password.";
-\! mkdir -p '{{backup_dir}}';
-\! sudo sudo mysqldump --add-drop-database --add-drop-trigger --log-error={{backup_dir}}/sql_error.log --databases '{{db_name}}' > '{{backup_dir}}/{{db_name}}.bak.sql';
+\! mkdir -p '#backup_dir#';
+\! sudo sudo mysqldump --add-drop-database --add-drop-trigger --log-error=#backup_dir#/sql_error.log --databases '#db_name#' > '#backup_dir#/#db_name#.bak.sql';
 \! echo "Backup completed succesfully...";
